requirements-nbwin.txt
absl-py==0.13.0
aiodns==2.0.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
anyio==3.2.1
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.3.4
astroid==2.6.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
black==21.6b0
bleach==3.3.0
blessings==1.7
blosc==1.10.4
boto3==1.4.8
botocore==1.8.50
cached-property==1.5.2
cachetools==4.2.2
catboost==0.26
ccxt==1.52.83
Cerberus==1.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
confi==0.0.4.1
configparser==5.0.2
coverage==5.5
coveralls==3.1.0
cryptography==3.4.7
cycler==0.10.0
dask==2021.8.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dfsql==0.4.3
dill==0.3.4
distlib==0.3.2
distributed==2021.8.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-to-markdown==0.9
docutils==0.17.1
entrypoints==0.3
fastapi==0.66.0
filelock==3.0.12
flake8==3.9.2
flake8-tidy-imports==4.3.0
flake8-type-annotations==0.1.0
fsspec==2021.7.0
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gpustat==0.6.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.39.0
h11==0.12.0
HeapDict==1.0.1
hiredis==2.0.0
idna==2.10
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.2
jedi==0.17.2
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-lsp==1.4.0
jupyter-resource-usage==0.6.0
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-lsp==3.8.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
locket==0.2.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mindsdb-sql==0.0.17
mistune==0.8.4
modin==0.10.1
modin-spreadsheet==0.1.2
msgpack==1.0.2
multidict==5.1.0
multitasking==0.0.9
mypy==0.910
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numexpr==2.7.3
numpy==1.21.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
orderedmultidict==1.0.1
packaging==20.9
pandas==1.3.0
pandas-ta==0.3.2b0
pandocfilters==1.4.3
parso==0.7.1
partd==1.2.0
pathspec==0.8.1
pathtools==0.1.2
pep517==0.10.0
pickleshare==0.7.5
Pillow==8.3.1
pip-api==0.0.20
pip-shims==0.5.3
pipreqs==0.4.10
plette==0.2.3
plotly==5.1.0
pluggy==0.13.1
plumbum==1.7.0
progressbar2==3.53.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
py-find-1st==1.1.5
py-spy==0.3.8
pyaml==20.4.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycodestyle==2.7.0
pycoingecko==2.2.0
pycparser==2.20
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.9.3
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-cov==2.12.1
pytest-mock==3.6.1
pytest-random-order==1.0.4
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2+49.g3536061
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.1.0
python-rapidjson==1.4
python-telegram-bot==13.7
python-utils==2.5.6
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
QuantStats==0.0.34
questionary==1.9.0
ray==1.5.2
redis==3.5.3
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requirementslib==1.5.16
retrying==1.3.3
rope==0.19.0
rpyc==4.1.5
rsa==4.7.2
s3transfer==0.1.13
scikit-learn==0.24.2
scikit-optimize==0.8.1
scipy==1.7.0
sdnotify==0.3.2
seaborn==0.11.1
Send2Trash==1.7.1
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
sly==0.4
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.20
stackprinter==0.2.5
starlette==0.14.2
subprocess32==3.5.4
TA-Lib @ file:///C:/CS/Python/crypto-collection/freqtrade-futures/build_helpers/TA_Lib-0.4.20-cp39-cp39-win_amd64.whl
tables @ file:///C:/CS/Python/crypto-collection/freqtrade-futures/build_helpers/tables-3.6.1-cp39-cp39-win_amd64.whl
tabulate==0.8.9
tblib==1.7.0
technical==1.3.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.7.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
types-cachetools==0.1.9
types-filelock==0.1.4
types-requests==2.25.0
types-tabulate==0.1.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.14.0
vistir==0.5.2
wandb==0.12.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
yarl==1.6.3
yfinance==0.1.62
zict==2.0.0
